fix(dashboard): 24h charts aligned relative to now
Fixes the timecharts displayed on the main dashboard to align the time
relative to now. Previously it aligned it to exact days.

This is done by a timechart argument `aligntime=latest`.
Time range of these searches is increased to -48 hours to cover the
values between 24h and 48 hours.
`earliest` search param is cleared out in favor of general search time
range setting.

EVNT-298